Refactor hacker (#357)
* init commit

* hacker refactor

* fixed description column error

* fixed error

* Run Prettier

---------
