+++ components/hacker/TeamSetup.tsx
@@ -2,9 +2,9 @@ import { Col, Divider, Row } from 'antd';
 import { Content } from 'antd/lib/layout/layout';
 import { useSWRConfig } from 'swr';
 import { ScopedMutator } from 'swr/dist/types';
-import { handleSubmitFailure } from '../lib/helpers';
-import { NewTeamFields } from '../types/client';
-import TeamCard from './TeamCard';
+import { handleSubmitFailure } from '../../lib/helpers';
+import { NewTeamFields } from '../../types/client';
+import TeamCard from '../TeamCard';
 
 async function handleSubmit(formData: NewTeamFields | { joinCode: string }, mutate: ScopedMutator<any>) {
 	const res = await fetch('/api/team-management', {
